% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlGLM.R
\name{dmlGLM}
\alias{dmlGLM}
\title{Double Machine Learning for Estimating Treatment Effects}
\usage{
dmlGLM(
  data = NULL,
  dep = NULL,
  treat = NULL,
  compile = TRUE,
  alpha = 1,
  nfolds = 10,
  splits = 2,
  DML = "DML1"
)
}
\arguments{
\item{data}{a data frame with dependent, treatment, and control variables.}

\item{dep}{the column number of the dependent variable.}

\item{treat}{the column number of the treatment variable.}

\item{compile}{if TRUE, summarize the results.}

\item{alpha}{what alpha to use for the internal glmnet. 1 = lasso, 0 = ridge.}

\item{nfolds}{number of folds to use in cross-validating nuisance function.}

\item{splits}{number of sample splits used to estimate treatment effect.}

\item{DML}{1 - theta as calculated in equation 1.3 of Chernozhukov et al. (2016).
2 - theta as calculated in the footnotes on page 4 of Chernozhukov et al. (2016).
3 - theta using the Frisch-Waugh-Lovell style residual on residual regression.}
}
\value{
A named vector or list
}
\description{
One of several functions in this package that performs the estimation of the
debiased estimator as outlined in Chernozhukov et al (2016). It requires the
user to provide a data frame, column indexes for the dependent and treatment
variable, how many splits to perform, and whether to summarize the results or
return all the output data in a list instead. The user also has a choice of
how theta is calculated.
}
\details{
dmlGLM specifically does this using glmnet as the foundational ML model for
fitting the nuisance function in cross-splitting.
}
\examples{
dmlGLM(data, dep = 1, treat = 2, splits = 3, DML = "FWL")
}
