% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlRFapply.R
\name{dmlRFapply}
\alias{dmlRFapply}
\title{Double Machine Learning for Esitmating Treatment Effects}
\usage{
dmlRFapply(
  data = NULL,
  dep = NULL,
  treat = NULL,
  compile = TRUE,
  splits = 2,
  DML = "DML1"
)
}
\arguments{
\item{data}{a data frame with depenent, treatment, and control variables.}

\item{dep}{the column number of the dependent variable.}

\item{treat}{the column number of the treatment variable.}

\item{compile}{if TRUE, summarize the results.}

\item{splits}{number of sample splits used to estimate treatment effect.}

\item{DML}{1 - theta as calculated in equation 1.3 of Chernozhukov et al. (2016).
2 - theta as calculated in the footnotes on page 4 of Chernozhukov et al. (2016).
3 - theta using the Frisch-Waugh-Lovell style residual on residual regression.}
}
\value{
A named vector or list
}
\description{
This function performs the estimation of the debiased estimator
as outlined in Chernozhukov et al (2016). It requres the user to
provide a data frame, column indexes for the depenent and treatment
variable, how many splits to perform, and whether to summarize
the results or return all the output data in a list instead. The user
also has a choice of how theta is calculated.
}
\details{
dmlRFapply differs from dmlRF in that the former fits the models to data
splits using the apply method while the latter does so by looping through the
splits.
}
\examples{
dmlRF(data, dep = 1, treat = 2, splits = 3, DML = "FWL")
}
